fix: Change uint to uint64 where necessary

+++ proc.h
@@ -38,7 +38,7 @@ enum procstate { UNUSED, EMBRYO, SLEEPING, RUNNABLE, RUNNING, ZOMBIE };
 
 // Per-process state
 struct proc {
-  uint sz;                     // Size of process memory (bytes)
+  uint64 sz;                   // Size of process memory (bytes)
   pml4e_t* pml4;               // Page table
   char *kstack;                // Bottom of kernel stack for this process
   enum procstate state;        // Process state
